+++ packages/web-platform/src/controllers/main.ts
@@ -1,59 +1,30 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Glue42Web } from "@glue42/web";
 import { UnsubscribeFunction } from "callback-registry";
-import { CoreClientData, InternalPlatformConfig, LibController, LibDomains } from "../common/types";
-import { libDomainDecoder } from "../shared/decoders";
+import { CoreClientData, InternalPlatformConfig } from "../common/types";
 import { GlueController } from "./glue";
-import { WindowsController } from "../libs/windows/controller";
 import { PortsBridge } from "../connection/portsBridge";
-import { ApplicationsController } from "../libs/applications/controller";
 import { WindowsStateController } from "./state";
 import logger from "../shared/logger";
 import { generate } from "shortid";
-import { LayoutsController } from "../libs/layouts/controller";
-import { WorkspacesController } from "../libs/workspaces/controller";
-import { IntentsController } from "../libs/intents/controller";
-import { ChannelsController } from "../libs/channels/controller";
 import { Glue42WebPlatform } from "../../platform";
-import { SystemController } from "./system";
 import { ServiceWorkerController } from "./serviceWorker";
-import { NotificationsController } from "../libs/notifications/controller";
-import { ExtensionController } from "../libs/extension/controller";
 import { PreferredConnectionController } from "../connection/preferred";
 import { Glue42Core } from "@glue42/core";
 import { InterceptionController } from "./interception";
 import { PluginsController } from "./plugins";
+import { DomainsController } from "./domains";
 
 export class PlatformController {
 
     private _platformApi!: Glue42WebPlatform.API;
 
-    private readonly controllers: { [key in LibDomains]: LibController } = {
-        system: this.systemController,
-        windows: this.windowsController,
-        appManager: this.applicationsController,
-        layouts: this.layoutsController,
-        workspaces: this.workspacesController,
-        intents: this.intentsController,
-        channels: this.channelsController,
-        notifications: this.notificationsController,
-        extension: this.extensionController
-    }
-
     constructor(
-        private readonly systemController: SystemController,
+        private readonly domainsController: DomainsController,
         private readonly glueController: GlueController,
-        private readonly windowsController: WindowsController,
-        private readonly applicationsController: ApplicationsController,
-        private readonly layoutsController: LayoutsController,
-        private readonly workspacesController: WorkspacesController,
-        private readonly intentsController: IntentsController,
-        private readonly channelsController: ChannelsController,
-        private readonly notificationsController: NotificationsController,
         private readonly portsBridge: PortsBridge,
         private readonly stateController: WindowsStateController,
         private readonly serviceWorkerController: ServiceWorkerController,
-        private readonly extensionController: ExtensionController,
         private readonly preferredConnectionController: PreferredConnectionController,
         private readonly interceptionController: InterceptionController,
         private readonly pluginsController: PluginsController
@@ -90,7 +61,7 @@ export class PlatformController {
 
         this.stateController.start();
 
-        await Promise.all(Object.values(this.controllers).map((controller) => controller.start(config)));
+        await this.domainsController.startAllDomains(config);
 
         await this.glueController.initClientGlue(config?.glue, config?.glueFactory, config?.workspaces?.isFrame);
 
@@ -98,14 +69,14 @@ export class PlatformController {
 
         this._platformApi = this.buildPlatformApi();
 
+        await this.pluginsController.startCorePlus(config);
+
         await this.pluginsController.start({
             platformConfig: config,
             plugins: config.plugins?.definitions,
             api: this.platformApi,
             handlePluginMessage: this.handlePluginMessage.bind(this)
         });
-
-        await this.pluginsController.startCorePlus(config.corePlus);
 
         if (config.connection.preferred) {
             await this.preferredConnectionController.start(config.connection.preferred);
@@ -131,24 +102,22 @@ export class PlatformController {
     }
 
     private async processControllerCommand(args: Glue42WebPlatform.Plugins.BaseControlMessage, callerType: "plugin" | "client", callerId: string): Promise<any> {
-        const decodeResult = libDomainDecoder.run(args.domain);
-
-        if (!decodeResult.ok) {
-            const errString = JSON.stringify(decodeResult.error);
+        try {
+            this.domainsController.validateDomain(args.domain);
+        } catch (error) {
+            const errString = JSON.stringify(error);
 
             this.logger?.trace(`rejecting execution of a command issued by a ${callerType}: ${callerId}, because of a domain validation error: ${errString}`);
 
             throw new Error(`Cannot execute this platform control, because of domain validation error: ${errString}`);
         }
 
-        const domain = decodeResult.result;
-
         const controlMessage: Glue42WebPlatform.ControlMessage = Object.assign({}, args, {
             commandId: generate(),
             callerId, callerType
         })
 
-        this.logger?.trace(`[${controlMessage.commandId}] received a command for a valid domain: ${domain} from ${callerType}: ${callerId}, forwarding to the appropriate controller`);
+        this.logger?.trace(`[${controlMessage.commandId}] received a command for a valid domain: ${args.domain} from ${callerType}: ${callerId}, forwarding to the appropriate controller`);
 
         try {
             const result = await this.executeCommand(controlMessage);
@@ -161,22 +130,12 @@ export class PlatformController {
 
             this.logger?.trace(`[${controlMessage.commandId}] this command's execution was rejected, reason: ${stringError}`);
 
-            throw new Error(`The platform rejected operation ${controlMessage.operation} for domain: ${domain} with reason: ${stringError}`);
+            throw new Error(`The platform rejected operation ${controlMessage.operation} for domain: ${args.domain} with reason: ${stringError}`);
         }
     }
 
     private handleClientUnloaded(client: CoreClientData): void {
-        this.logger?.trace(`detected unloading of client: ${client.windowId}, notifying all controllers`);
-
-        Object.values(this.controllers).forEach((controller, idx) => {
-            try {
-                controller.handleClientUnloaded?.(client.windowId, client.win);
-            } catch (error: any) {
-                const stringError = typeof error === "string" ? error : JSON.stringify(error.message);
-                const controllerName = Object.keys(this.controllers)[idx];
-                this.logger?.error(`${controllerName} controller threw when handling unloaded client ${client.windowId} with error message: ${stringError}`);
-            }
-        });
+        this.domainsController.notifyDomainsClientUnloaded(client);
     }
 
     private executeCommand(controlMessage: Glue42WebPlatform.ControlMessage): Promise<any> {
@@ -188,7 +147,7 @@ export class PlatformController {
             return interceptor.intercept(controlMessage);
         }
 
-        return this.controllers[controlMessage.domain].handleControl(controlMessage);
+        return this.domainsController.executeControlMessage(controlMessage);
     }
 
     private buildPlatformApi(): Glue42WebPlatform.API {